Added all the variations of move anims to test
